# internal/user/api/language.go
package api

import (
	"net/http"

	"github.com/econumo/econumo/internal/model"
	"github.com/econumo/econumo/internal/web/apidoc"
	"github.com/econumo/econumo/internal/web/endpoint"
)

// _ keeps the apidoc/model import aliases visible to swag's annotation
// parser (this file's handler body no longer references model types
// directly, since it delegates to a method value).
var _ = apidoc.JsonResponseError{}
var _ = model.UpdateLanguageResult{}

// UpdateLanguage handles POST /api/v1/user/update-language (auth). The JSON
// field is "language"; tier-1 validates NotBlank, the service enforces
// membership in the supported-language set.
//
// @Summary     Update language
// @Description Updates the authenticated user's UI language preference and returns the refreshed user.
// @Tags        User
// @Accept      json
// @Produce     json
// @Param       request body     model.UpdateLanguageRequest true "Update language request"
// @Success     200     {object} apidoc.JsonResponseOk{data=model.UpdateLanguageResult}
// @Failure     400     {object} apidoc.JsonResponseError
// @Failure     401     {object} apidoc.JsonResponseUnauthorized
// @Failure     402     {object} apidoc.JsonResponseError
// @Failure     500     {object} apidoc.JsonResponseException
// @Security    Bearer
// @Router      /api/v1/user/update-language [post]
func (h *Handlers) UpdateLanguage(w http.ResponseWriter, r *http.Request) {
	endpoint.Handle(w, r, h.svc.UpdateLanguage)
}