# internal/web/apidoc/docs/swagger.yaml
basePath: /
definitions:
  apidoc.JsonResponseError:
    properties:
      code:
        example: 0
        type: integer
      errors:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      message:
        example: Validation failed
        type: string
      success:
        example: false
        type: boolean
    type: object
  apidoc.JsonResponseException:
    properties:
      code:
        example: 0
        type: integer
      exceptionType:
        type: string
      message:
        example: Internal Server Error
        type: string
      success:
        example: false
        type: boolean
    type: object
  apidoc.JsonResponseOk:
    properties:
      data: {}
      message:
        example: ""
        type: string
      success:
        example: true
        type: boolean
    type: object
  apidoc.JsonResponseUnauthorized:
    properties:
      code:
        example: 0
        type: integer
      errors:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      message:
        example: Access token not found
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  model.ConfirmEmailRequest:
    properties:
      code:
        type: string
      username:
        type: string
    type: object
  model.ConfirmEmailResult:
    type: object
...
  model.CurrentUserResult:
    properties:
      accessLevel:
        type: string
      accessUntil:
        type: string
      avatar:
        type: string
      currency:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      options:
        items:
          $ref: '#/definitions/model.OptionResult'
        type: array
      reportPeriod:
        type: string
    type: object
...
  model.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  model.LoginResult:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/model.CurrentUserResult'
    type: object
  model.LogoutResult:
    properties:
      result:
        type: string
    type: object
...
  model.OptionResult:
    properties:
      name:
        type: string
      value:
        type: string
    type: object
...
  model.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  model.RemindPasswordRequest:
    properties:
      username:
        type: string
    type: object
  model.RemindPasswordResult:
    type: object
...
  model.ResendVerificationCodeRequest:
    properties:
      username:
        type: string
    type: object
  model.ResendVerificationCodeResult:
    type: object
...
  model.ResetPasswordRequest:
    properties:
      code:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  model.ResetPasswordResult:
    type: object
...
  model.UpdateLanguageRequest:
    properties:
      language:
        type: string
    type: object
  model.UpdateLanguageResult:
    properties:
      user:
        $ref: '#/definitions/model.CurrentUserResult'
    type: object
...
  model.UpdatePasswordRequest:
    properties:
      newPassword:
        type: string
      oldPassword:
        type: string
    type: object
  model.UpdatePasswordResult:
    type: object
...
info:
  contact: {}
  description: Self-hosted personal finance and budgeting API
  title: Econumo API
  version: 1.0.0
paths:
...
  /api/v1/user/confirm-email:
    post:
      consumes:
      - application/json
      description: Confirms a user's email with the emailed verification code (ECONUMO_EMAIL_VERIFICATION).
        Returns an empty success envelope.
      parameters:
      - description: Confirm email request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ConfirmEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.ConfirmEmailResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.JsonResponseException'
      summary: Confirm email
      tags:
      - User
...
  /api/v1/user/login-user:
    post:
      consumes:
      - application/json
      description: Authenticates a user by username/password and returns an opaque
        access token plus the current user.
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Raw {token,user} body — NOT wrapped in the standard envelope
            (matches PHP login).
          schema:
            $ref: '#/definitions/model.LoginResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.JsonResponseUnauthorized'
        "403":
          description: 'Email verification required (ECONUMO_EMAIL_VERIFICATION):
            confirm via /api/v1/user/confirm-email, then log in again.'
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.JsonResponseException'
      summary: Log in
      tags:
      - User
  /api/v1/user/logout-user:
    post:
      description: Revokes the presenting session token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.LogoutResult'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.JsonResponseUnauthorized'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.JsonResponseException'
      security:
      - Bearer: []
      summary: Log out
      tags:
      - User
  /api/v1/user/register-user:
    post:
      consumes:
      - application/json
      description: Creates a new user (when registration is enabled) and returns the
        current user without a token.
      parameters:
      - description: Register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.CurrentUserResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.JsonResponseUnauthorized'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.JsonResponseException'
      summary: Register a user
      tags:
      - User
  /api/v1/user/remind-password:
    post:
      consumes:
      - application/json
      description: Triggers a password reminder email. Always returns success (anti-enumeration).
      parameters:
      - description: Remind password request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.RemindPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.RemindPasswordResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.JsonResponseUnauthorized'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.JsonResponseException'
      summary: Remind password
      tags:
      - User
...
  /api/v1/user/resend-verification-code:
    post:
      consumes:
      - application/json
      description: Re-sends the email verification code, at most once per 60s. Always
        returns success (anti-enumeration); the Retry-After header carries the seconds
        until another code may be requested.
      parameters:
      - description: Resend verification code request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ResendVerificationCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'Retry-After: seconds until another code may be requested'
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.ResendVerificationCodeResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.JsonResponseException'
      summary: Resend verification code
      tags:
      - User
  /api/v1/user/reset-password:
    post:
      consumes:
      - application/json
      description: Resets a user's password using a reminder code. Returns an empty
        success envelope.
      parameters:
      - description: Reset password request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.ResetPasswordResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.JsonResponseUnauthorized'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.JsonResponseException'
      summary: Reset password
      tags:
      - User
...
  /api/v1/user/update-language:
    post:
      consumes:
      - application/json
      description: Updates the authenticated user's UI language preference and returns
        the refreshed user.
      parameters:
      - description: Update language request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateLanguageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.UpdateLanguageResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.JsonResponseUnauthorized'
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.JsonResponseException'
      security:
      - Bearer: []
      summary: Update language
      tags:
      - User
...
  /api/v1/user/update-password:
    post:
      consumes:
      - application/json
      description: Changes the authenticated user's password after verifying the old
        one; other sessions are signed out. Returns an empty success envelope.
      parameters:
      - description: Update password request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apidoc.JsonResponseOk'
            - properties:
                data:
                  $ref: '#/definitions/model.UpdatePasswordResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apidoc.JsonResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apidoc.JsonResponseUnauthorized'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apidoc.JsonResponseException'
      security:
      - Bearer: []
      summary: Update password
      tags:
      - User
...
securityDefinitions:
  Bearer:
    description: 'Opaque access token. Format: "Bearer <token>".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"